refactor(app): the rest of em
