+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -25,9 +25,9 @@ import { EstopTakeover } from '../organisms/EmergencyStop'
 import { ConnectViaEthernet } from '../pages/ODD/ConnectViaEthernet'
 import { ConnectViaUSB } from '../pages/ODD/ConnectViaUSB'
 import { ConnectViaWifi } from '../pages/ODD/ConnectViaWifi'
-import { EmergencyStop } from '../pages/EmergencyStop'
-import { NameRobot } from '../pages/NameRobot'
-import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
+import { EmergencyStop } from '../pages/ODD/EmergencyStop'
+import { NameRobot } from '../pages/ODD/NameRobot'
+import { NetworkSetupMenu } from '../pages/ODD/NetworkSetupMenu'
 import { ProtocolSetup } from '../pages/ODD/ProtocolSetup'
 import { RobotDashboard } from '../pages/ODD/RobotDashboard'
 import { RobotSettingsDashboard } from '../pages/ODD/RobotSettingsDashboard'
@@ -42,9 +42,9 @@ import { UpdateRobot } from '../pages/ODD/UpdateRobot/UpdateRobot'
 import { UpdateRobotDuringOnboarding } from '../pages/ODD/UpdateRobot/UpdateRobotDuringOnboarding'
 import { InstrumentsDashboard } from '../pages/ODD/InstrumentsDashboard'
 import { InstrumentDetail } from '../pages/ODD/InstrumentDetail'
-import { Welcome } from '../pages/Welcome'
-import { InitialLoadingScreen } from '../pages/InitialLoadingScreen'
-import { DeckConfigurationEditor } from '../pages/DeckConfiguration'
+import { Welcome } from '../pages/ODD/Welcome'
+import { InitialLoadingScreen } from '../pages/ODD/InitialLoadingScreen'
+import { DeckConfigurationEditor } from '../pages/ODD/DeckConfiguration'
 import { PortalRoot as ModalPortalRoot } from './portal'
 import { getOnDeviceDisplaySettings, updateConfigValue } from '../redux/config'
 import { updateBrightness } from '../redux/shell'
