+++ app/src/pages/ODD/EmergencyStop/__tests__/EmergencyStop.test.tsx
@@ -3,9 +3,9 @@ import { vi, it, describe, expect, beforeEach } from 'vitest'
 import { useEstopQuery } from '@opentrons/react-api-client'
 import { fireEvent, screen } from '@testing-library/react'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { EmergencyStop } from '..'
 import type { NavigateFunction } from 'react-router-dom'
