+++ app/src/pages/ODD/InitialLoadingScreen/__tests__/InitialLoadingScreen.test.tsx
@@ -4,9 +4,9 @@ import { screen } from '@testing-library/react'
 
 import { useRobotSettingsQuery } from '@opentrons/react-api-client'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 
-import { getIsShellReady } from '../../../redux/shell'
+import { getIsShellReady } from '../../../../redux/shell'
 
 import { InitialLoadingScreen } from '..'
 
@@ -14,8 +14,8 @@ import type { UseQueryResult } from 'react-query'
 import type { RobotSettingsResponse } from '@opentrons/api-client'
 
 vi.mock('@opentrons/react-api-client')
-vi.mock('../../../redux/config')
-vi.mock('../../../redux/shell')
+vi.mock('../../../../redux/config')
+vi.mock('../../../../redux/shell')
 
 const render = () => {
   return renderWithProviders(<InitialLoadingScreen />)
