+++ app/src/pages/ODD/NetworkSetupMenu/__tests__/NetworkSetupMenu.test.tsx
@@ -2,9 +2,9 @@ import * as React from 'react'
 import { vi, it, describe, expect } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { NetworkSetupMenu } from '..'
 import type { NavigateFunction } from 'react-router-dom'
