+++ app/src/pages/ODD/NetworkSetupMenu/index.tsx
@@ -13,8 +13,8 @@ import {
   TYPOGRAPHY,
 } from '@opentrons/components'
 
-import { StepMeter } from '../../atoms/StepMeter'
-import { CardButton } from '../../molecules/CardButton'
+import { StepMeter } from '../../../atoms/StepMeter'
+import { CardButton } from '../../../molecules/CardButton'
 
 import type { IconName } from '@opentrons/components'
