+++ app/src/pages/ODD/Welcome/__tests__/Welcome.test.tsx
@@ -3,9 +3,9 @@ import { vi, it, describe, expect } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 
-import { i18n } from '../../../i18n'
+import { i18n } from '../../../../i18n'
 import { Welcome } from '..'
 
 import type { NavigateFunction } from 'react-router-dom'
