+++ app/src/pages/ODD/Welcome/index.tsx
@@ -10,9 +10,9 @@ import {
   LegacyStyledText,
   TYPOGRAPHY,
 } from '@opentrons/components'
-import { MediumButton } from '../../atoms/buttons'
+import { MediumButton } from '../../../atoms/buttons'
 
-import screenImage from '../../assets/images/on-device-display/welcome_background.png'
+import screenImage from '../../../assets/images/on-device-display/welcome_background.png'
 
 const IMAGE_ALT = 'Welcome screen background image'
